requirements.txt
alabaster==0.7.9
anaconda-client==1.2.2
appdirs==1.4.3
appnope==0.1.0
args==0.1.0
astroid==1.4.9
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0rc1
backports.weakref==1.0rc1
beautifulsoup4==4.5.3
bitarray==0.8.1
blaze==0.10.1
blessed==1.14.2
bokeh==0.12.4
boto==2.45.0
botocore==1.5.80
Bottleneck==1.1.0
cement==2.8.2
certifi==2017.4.17
chardet==3.0.4
chest==0.2.3
click==6.7
clint==0.5.1
cloudpickle==0.2.2
clyent==1.2.1
colorama==0.3.7
conda==4.3.16
configobj==5.0.6
contextlib2==0.5.4
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.13.0
datashape==0.5.2
decorator==4.0.11
dill==0.2.5
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
et-xmlfile==1.0.1
fastcache==1.0.2
Flask~> 0.12.3
Flask-Cors==3.0.2
floyd-cli==0.9.5
gevent==1.2.1
greenlet==0.4.11
gunicorn==19.7.1
HeapDict==1.0.0
idna==2.5
imagesize==0.7.1
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.9.0
Jinja2==2.9.4
jmespath==0.9.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==2.0.4
lazy-object-proxy==1.2.2
locket==0.2.0
lxml==3.7.2
MarkupSafe==0.23
marshmallow ~> 2.15.1.
mistune==0.7.3
mpmath==0.19
multipledispatch==0.4.9
nbconvert==4.2.0
nbformat==4.2.0
networkx==1.11
nltk==3.2.4
nose==1.3.7
notebook==5.4.1
numpy==1.12.1
numpydoc==0.6.0
odo==0.5.0
olefile==0.44
opencv-python
openpyxl==2.4.1
pandas==0.19.2
packaging==16.8
partd==0.3.7
pathlib2==2.2.1
pathspec==0.5.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.3.0
psutil==5.0.1
ptyprocess==0.5.1
py==1.4.32
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.17
pyflakes==1.5.0
Pygments==2.1.3
pylint==1.6.4
pyopenssl ~> 17.5.0
pyparsing==2.2.0
pytest==3.0.5
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.5
requests~>2.20.0
requests-toolbelt==0.8.0
rope-py3k==0.9.4.post1
scikit-learn==0.18.1
semantic-version==2.5.0
shortuuid==0.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.5.1
SQLAlchemy==1.1.5
sympy==1.0
tensorflow==1.0.0
termcolor==1.1.0
terminado==0.6
toolz==0.8.2
tornado==4.4.2
traitlets==4.3.1
unicodecsv==0.14.1
urllib3==1.21.1
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.12.2
widgetsnbextension==1.2.6
wordcloud==1.3.1
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
